+++ editors/code/src/command/workspace-folder-formatting.ts
@@ -3,15 +3,13 @@ import * as vscode from "vscode";
 import { Cmd, Ctx } from "../context";
 import * as output from "../output";
 import { isError, isResult, runCommand } from "../process";
+import {
+	saveAllDirtyWorkspaceTextDocuments,
+	selectWorkspaceFolder,
+} from "../workspace";
 
 /**
  * Format a workspace folder
- *
- * # Workspace folder selection
- *
- * - If 0 workspace folders are open, errors
- * - If 1 workspace folder is open, automatically uses it
- * - If >1 workspace folders are open, asks the user to choose
  *
  * # Tab saving
  *
@@ -42,195 +40,78 @@ import { isError, isResult, runCommand } from "../process";
  * `{root}/R/air.toml` that would handle any R files in your project. Instead,
  * we hope this isn't common enough to come up much in practice.
  */
-export function workspaceFolderFormatting(ctx: Ctx): Cmd {
-	return async () => {
-		const binaryPath = ctx.lsp.getBinaryPath();
+export async function workspaceFolderFormatting(
+	workspaceFolder: vscode.WorkspaceFolder,
+	binaryPath: string,
+) {
+	const allSaved = await saveAllDirtyWorkspaceTextDocuments(workspaceFolder);
+	if (!allSaved) {
+		return;
+	}
 
+	const workspaceFolderPath = workspaceFolder.uri.fsPath;
+
+	// i.e., `air format {workspaceFolderPath} --no-color`
+	const args = ["format", workspaceFolderPath, "--no-color"];
+
+	// This should not matter since the path is explicitly supplied, but better to be safe
+	const options = {
+		cwd: workspaceFolderPath,
+	};
+
+	// Resolves when the spawned process closes or errors
+	const result = await runCommand(binaryPath, args, options);
+
+	let anyErrors = false;
+
+	if (isError(result)) {
+		// Something went horribly wrong in the process spawning or shutdown process
+		anyErrors = true;
+		output.log(
+			`Errors occurred while formatting the ${workspaceFolder.name} workspace folder.\n${result.error.message}`,
+		);
+	}
+
+	if (isResult(result)) {
+		if (result.code !== 0) {
+			// Air was able to run and exit, but we had an error along the way
+			output.log(
+				`Errors occurred while formatting the ${workspaceFolder.name} workspace folder.\n${result.stderr}`,
+			);
+			anyErrors = true;
+		}
+	}
+
+	if (anyErrors) {
+		const answer = await vscode.window.showInformationMessage(
+			`Errors occurred while formatting the ${workspaceFolder.name} workspace folder. View the logs?`,
+			{ modal: true },
+			"Yes",
+			"No",
+		);
+
+		if (answer === "Yes") {
+			output.show();
+		}
+
+		return;
+	}
+
+	vscode.window.showInformationMessage(
+		`Successfully formatted the ${workspaceFolder.name} workspace folder.`,
+	);
+}
+
+export function workspaceFolderFormattingCallback(ctx: Ctx): Cmd {
+	return async () => {
 		const workspaceFolder = await selectWorkspaceFolder();
+
 		if (!workspaceFolder) {
 			return;
 		}
 
-		const allSaved =
-			await saveAllDirtyWorkspaceTextDocuments(workspaceFolder);
-		if (!allSaved) {
-			return;
-		}
+		const binaryPath = ctx.lsp.getBinaryPath();
 
-		const workspaceFolderPath = workspaceFolder.uri.fsPath;
-
-		// i.e., `air format {workspaceFolderPath} --no-color`
-		const args = ["format", workspaceFolderPath, "--no-color"];
-
-		// This should not matter since the path is explicitly supplied, but better to be safe
-		const options = {
-			cwd: workspaceFolderPath,
-		};
-
-		// Resolves when the spawned process closes or errors
-		const result = await runCommand(binaryPath, args, options);
-
-		let anyErrors = false;
-
-		if (isError(result)) {
-			// Something went horribly wrong in the process spawning or shutdown process
-			anyErrors = true;
-			output.log(
-				`Errors occurred while formatting the ${workspaceFolder.name} workspace folder.\n${result.error.message}`,
-			);
-		}
-
-		if (isResult(result)) {
-			if (result.code !== 0) {
-				// Air was able to run and exit, but we had an error along the way
-				output.log(
-					`Errors occurred while formatting the ${workspaceFolder.name} workspace folder.\n${result.stderr}`,
-				);
-				anyErrors = true;
-			}
-		}
-
-		if (anyErrors) {
-			const answer = await vscode.window.showInformationMessage(
-				`Errors occurred while formatting the ${workspaceFolder.name} workspace folder. View the logs?`,
-				{ modal: true },
-				"Yes",
-				"No",
-			);
-
-			if (answer === "Yes") {
-				output.show();
-			}
-
-			return;
-		}
-
-		vscode.window.showInformationMessage(
-			`Successfully formatted the ${workspaceFolder.name} workspace folder.`,
-		);
+		await workspaceFolderFormatting(workspaceFolder, binaryPath);
 	};
-}
-
-async function selectWorkspaceFolder(): Promise<
-	vscode.WorkspaceFolder | undefined
-> {
-	const workspaceFolders = vscode.workspace.workspaceFolders;
-
-	if (!workspaceFolders || workspaceFolders.length === 0) {
-		vscode.window.showErrorMessage(
-			"You must be inside a workspace to format a workspace folder.",
-		);
-		return undefined;
-	}
-
-	if (workspaceFolders.length === 1) {
-		return workspaceFolders[0];
-	}
-
-	// Let the user select a workspace folder if >1 are open, may be
-	// `undefined` if user bails from quick pick!
-	const workspaceFolder =
-		await selectWorkspaceFolderFromQuickPick(workspaceFolders);
-
-	return workspaceFolder;
-}
-
-async function selectWorkspaceFolderFromQuickPick(
-	workspaceFolders: readonly vscode.WorkspaceFolder[],
-): Promise<vscode.WorkspaceFolder | undefined> {
-	// Show the workspace names
-	const workspaceFolderNames = workspaceFolders.map(
-		(workspaceFolder) => workspaceFolder.name,
-	);
-
-	const workspaceFolderName = await vscode.window.showQuickPick(
-		workspaceFolderNames,
-		{
-			canPickMany: false,
-			title: "Which workspace folder should be formatted?",
-		},
-	);
-
-	if (!workspaceFolderName) {
-		// User bailed from the quick pick
-		return undefined;
-	}
-
-	// Match selected name back to the workspace folder
-	for (let workspaceFolder of workspaceFolders) {
-		if (workspaceFolder.name === workspaceFolderName) {
-			return workspaceFolder;
-		}
-	}
-
-	// Should never get here
-	output.log(
-		`Matched a workspace folder name, but unexpectedly can't find corresponding workspace folder. Folder name: ${workspaceFolderName}.`,
-	);
-	return undefined;
-}
-
-/**
- * Save all open dirty editor tabs relevant to the workspace folder
- *
- * - Filters to only tabs living under the chosen workspace folder
- * - Asks the user if they are okay with us saving the editor tabs
- */
-async function saveAllDirtyWorkspaceTextDocuments(
-	workspaceFolder: vscode.WorkspaceFolder,
-): Promise<boolean> {
-	const textDocuments = dirtyWorkspaceTextDocuments(workspaceFolder);
-
-	if (textDocuments.length === 0) {
-		// Nothing to save!
-		return true;
-	}
-
-	// Ask the user if we can save them
-	const answer = await vscode.window.showInformationMessage(
-		`All editors within the ${workspaceFolder.name} workspace folder must be saved before formatting. Proceed with saving these editors?`,
-		{ modal: true },
-		"Yes",
-		"No",
-	);
-
-	if (answer !== "Yes") {
-		// User said `"No"` or bailed from the menu
-		return false;
-	}
-
-	// Save all documents, and ensure that all successfully saved
-	const savedPromises = textDocuments.map((textDocument) =>
-		textDocument.save(),
-	);
-	const saved = await Promise.all(savedPromises);
-	return saved.every((save) => save);
-}
-
-function dirtyWorkspaceTextDocuments(
-	workspaceFolder: vscode.WorkspaceFolder,
-): vscode.TextDocument[] {
-	return vscode.workspace.textDocuments.filter((document) => {
-		if (document.isClosed) {
-			// Not actually synchonized. This document will be refreshed when the document is reopened.
-			return false;
-		}
-
-		if (!document.isDirty) {
-			// Nothing to do
-			return false;
-		}
-
-		if (document.isUntitled) {
-			// These aren't part of the workspace folder
-			return false;
-		}
-
-		if (!document.uri.fsPath.startsWith(workspaceFolder.uri.fsPath)) {
-			// The document must live "under" the chosen workspace folder for us to care about it
-			return false;
-		}
-
-		return true;
-	});
 }